updates based on copilot feedback
